Coloca as similaridades na arvore AVL

+++ Resultado.java
@@ -1,9 +1,9 @@
 public class Resultado implements Comparable<Resultado> {
     private String arquivo1;
     private String arquivo2;
-    private float similaridade;
+    private double similaridade;
 
-    public Resultado(String arquivo1, String arquivo2, float similaridade) {
+    public Resultado(String arquivo1, String arquivo2, double similaridade) {
         this.arquivo1 = arquivo1;
         this.arquivo2 = arquivo2;
         this.similaridade = similaridade;
@@ -11,12 +11,12 @@ public class Resultado implements Comparable<Resultado> {
 
     public String getArquivo1() { return arquivo1; }
     public String getArquivo2() { return arquivo2; }
-    public float getSimilaridade() { return similaridade; }
+    public double getSimilaridade() { return similaridade; }
 
     
     @Override
     public int compareTo(Resultado outro) {
-        return Float.compare(this.similaridade, outro.similaridade);
+        return Double.compare(this.similaridade, outro.similaridade);
     }
 
     @Override
